+++ cognite/client/_api/datapoints.py
@@ -1,4 +1,5 @@
 import math
+import re
 import threading
 from collections import defaultdict, namedtuple
 from datetime import datetime
@@ -367,6 +368,8 @@ class DatapointsAPI(APIClient):
             str, List[str], Dict[str, Union[int, List[str]]], List[Dict[str, Union[int, List[str]]]]
         ] = None,
         limit: int = None,
+        complete=False,
+        dropna=False,
     ):
         """Get a pandas dataframe describing the requested data.
 
@@ -382,7 +385,10 @@ class DatapointsAPI(APIClient):
             external_id (Union[str, List[str], Dict[str, Any], List[Dict[str, Any]]]): External id or list of external
                 ids. Can also be object specifying aggregates. See example below.
             limit (int): Maximum number of datapoints to return for each time series.
-
+            complete (bool): insert missing entries into the index, and complete data where possible (supports interpolation, stepInterpolation, count, sum, totalVariation)
+            dropna (bool): Sets complete=True and drop rows in which any aggregate for any time series has missing values (typically rows at the start and end for interpolation aggregates).
+                    This guarantees that all returned dataframes have the exact same shape and no missing values anywhere.
+                    only supported for aggregates using any of sum, count, totalVariance, interpolation and stepInterpolation
         Returns:
             pandas.DataFrame: The requested dataframe
 
@@ -411,7 +417,88 @@ class DatapointsAPI(APIClient):
                 granularity=granularity,
                 limit=limit,
             ).to_pandas()
-        return pd.concat([id_df, external_id_df], axis="columns")
+        df = pd.concat([id_df, external_id_df], axis="columns")
+
+        if complete == "fill" or complete == "dropna":
+            if "interpolation" in aggregates:
+                id_list = id if id is None or isinstance(id, list) else [id]
+                eid_list = external_id if external_id is None or isinstance(external_id, list) else [external_id]
+                ts_meta = self._cognite_client.time_series.retrieve_multiple(ids=id_list, external_ids=eid_list)
+                is_step_dict = {
+                    **{ts.id: bool(ts.is_step) for ts in ts_meta},
+                    **{ts.external_id: bool(ts.is_step) for ts in ts_meta},
+                }
+            else:
+                is_step_dict = None
+            df = self._complete_dataframe(df, granularity, is_step_dict)
+        elif complete:
+            raise ValueError("complete should be 'fill', 'dropna' or Falsy")
+
+        if complete == "dropna":
+            completed_aggregates = ["sum", "count", "totalVariance", "interpolation", "stepInterpolation"]
+            if set(aggregates) - set(completed_aggregates):
+                raise ValueError(
+                    "One of the aggregates {} not in the supported aggregates {} for dropna".format(
+                        complete, completed_aggregates
+                    )
+                )
+            df.dropna(inplace=True)
+
+        return df
+
+    def retrieve_dataframe_dict(
+        self,
+        start: Union[int, str, datetime],
+        end: Union[int, str, datetime],
+        aggregates: List[str],
+        granularity: str,
+        id: Union[int, List[int], Dict[str, Union[int, List[str]]], List[Dict[str, Union[int, List[str]]]]] = None,
+        external_id: Union[
+            str, List[str], Dict[str, Union[int, List[str]]], List[Dict[str, Union[int, List[str]]]]
+        ] = None,
+        limit: int = None,
+        complete=None,
+    ):
+        """Get a dictionary of aggregate: pandas dataframe describing the requested data.
+
+        Args:
+
+            For other arguments, see retrieve_dataframe
+
+        Returns:
+            Union[pandas.DataFrame,Dict[pandas.DataFrame]]: The requested dataframe if one aggregate was requested, or a dictionary of aggregate: dataframe if multiple aggregates were requested.
+        """
+        df = self.retrieve_dataframe(start, end, aggregates, granularity, id, external_id, limit, complete)
+        df_dict = {ag: df.filter(like="|" + ag) for ag in aggregates}
+        return df_dict[aggregates[0]] if len(aggregates) == 1 else df_dict
+
+    def _complete_dataframe(self, df, granularity, is_step_dict):
+        if df.empty:
+            return df
+        np, pd = utils._auxiliary.local_import("numpy", "pandas")
+        df = df.reindex(
+            np.arange(
+                df.index[0],
+                df.index[-1] + pd.Timedelta(microseconds=1),
+                pd.Timedelta(microseconds=cognite.client.utils._time.granularity_to_ms(granularity) * 1000),
+            ),
+            copy=False,
+        )
+        df.fillna({c: 0 for c in df.columns if re.search(c, r"\|(sum|totalVariance|count)$")}, inplace=True)
+        # these can't be done in-place
+        if is_step_dict:  # iff interpolation in aggregates
+            int_cols = [c for c in df.columns if re.search(c, r"\|(interpolation)$")]
+            lin_int_cols = [c for c in int_cols if not is_step_dict[c[:-14]]]  # 14 is length of '|interpolation'
+            stepint_cols = list(np.setdiff1d(int_cols, lin_int_cols))
+            if lin_int_cols:
+                df[lin_int_cols] = df[lin_int_cols].interpolate()
+        else:
+            stepint_cols = []
+        stepint_cols += [c for c in df.columns if re.search(c, r"\|(stepInterpolation)$")]
+        if stepint_cols:
+            df[stepint_cols] = df[stepint_cols].ffill()
+
+        return df
 
     def insert_dataframe(self, dataframe, external_id_headers: bool = False):
         """Insert a dataframe.